+++ R/call.R
@@ -2,25 +2,22 @@
 # given a call of the form e.g. 'pkg::foo()' or 'foo()',
 # check that method 'foo()' is truly being called and
 # strip off the 'pkg::' part for easier parsing.
-#
-# this gets called very often when parsing dependencies,
-# so optimizations are welcome here
 renv_call_expect <- function(node, package, methods) {
 
   result <- NULL
-  
+
   # check for call of the form 'pkg::foo(a, b, c)'
   if (is.call(call <- node[[1L]]))
     if (is.symbol(symbol <- call[[1L]]))
       if (symbol == "::" || symbol == ":::")
         if (call[[2L]] == package)
           node[[1L]] <- call[[3L]]
-  
+
   # check for any method match
   if (is.symbol(symbol <- node[[1L]]))
     if (any(symbol == methods))
       result <- node
-  
+
   result
 
 }
@@ -65,12 +62,12 @@ renv_call_normalize <- function(node) {
 renv_call_matches <- function(call, names, nargs = NULL) {
 
   ok <- FALSE
-  
+
   if (is.call(call))
     if (is.symbol(sym <- call[[1L]]))
       if (any(names == sym))
         ok <- is.null(nargs) || length(call) == nargs + 1L
-  
+
   ok
-  
+
 }