+++ tests/testthat/resources/geom-hex.R
@@ -1,0 +1,4 @@
+library(ggplot2)
+
+d <- ggplot(diamonds, aes(carat, price))
+d + geom_hex()