+++ tests/testthat/resources/use.R
@@ -1,0 +1,10 @@
+
+# base R usages
+use("A")
+base::use("B", c("filter", "select"))
+
+# renv style usages
+renv::use(
+  C = "tidyverse/dplyr",
+  D = "r-lib/rlang"
+)