+++ NEWS.md
@@ -1,4 +1,6 @@
 
+# renv (development version)
+
 # renv 1.1.4
 
 * Fixed an issue where `renv` could erroneously record packages installed